+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Added missing `/status` subresource permission to RBAC role
+
 ## [1.1.0] - 2022-09-23
 
 ### Added